sched/tune: move schedtune_nornalize_energy into fair.c
The energy normalization function is required to get the proper values
for the P-E space filtering function to work.
That normalization is part of the hot wakeup path and currently implemented
with a function call.

Moving the normalization function into fair.c allows the compiler to
further optimize that code by reducing overheads in the wakeup hot path.

Change-Id: I86b6793e170b973e08819a6046c205a1060ab922
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: shagbag913 <sh4gbag913@gmail.com>